Renames pipeValue -> pipe
