+++ tests/pipe.test.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 
-import pipeValue from '../source/pipe-value';
+import pipe from '../source/pipe';
 
 // These tests are a bit contrived since many of the native things can already be chained.
 
@@ -14,11 +14,34 @@ const template = (t, input, expected) => {
   t.is(expected, input);
 };
 
-test('1', template, pipeValue('h e l l o', capitalize, removeSpaces), 'HELLO');
-test('2', template, pipeValue('a1b2c3', removeNonNumbers, Number), 123);
+test(
+  '1',
+  template,
+  pipe(
+    'h e l l o',
+    capitalize,
+    removeSpaces
+  ),
+  'HELLO'
+);
+test(
+  '2',
+  template,
+  pipe(
+    'a1b2c3',
+    removeNonNumbers,
+    Number
+  ),
+  123
+);
 test(
   '3',
   template,
-  pipeValue({ a: 1, b: 2 }, Object.entries, map(tuple => tuple[1]), sum),
+  pipe(
+    { a: 1, b: 2 },
+    Object.entries,
+    map(tuple => tuple[1]),
+    sum
+  ),
   3
 );
